requirements.txt
flask
diffusers
torch
pillow
numpy
gunicorn
safetensors
accelerate
transformers
xformers